+++ AntIndex/Models/Runtime/Requests/Search.cs
@@ -21,7 +21,7 @@ public class Search(
     public override void ProcessRequest(
         AntHill index,
         SearchContextBase searchContext,
-        Dictionary<int, ushort>[] wordsBundle,
+        Dictionary<int, byte>[] wordsBundle,
         CancellationToken ct)
     {
         if (!index.Entities.TryGetValue(EntityType, out var entities))
